+++ NEWS.md
@@ -21,7 +21,7 @@ Everything denoted by “formerly” refers to the previous name in [`Manifolds.
   * `SpecialOrthogonalGroup` (formerly `SpecialOrthogonal`)
   * `SpecialUnitaryGroup` (formerly `SpecialUnitary`)
   * `OrthogonalGroup` (formerly `Orthogonal`)
-  * `UnitaryGroup` (formerly `Unitary`) als for quaternions.
+  * `UnitaryGroup` (formerly `Unitary`) also for quaternions.
   * `GeneralLinearGroup` (formerly `GeneralLinear`)
   * `HeisenbergGroup`
   * `LeftSemidirectProductLieGroup` (formerly `SemidirectProductGroup`)